diffraction metadata presenter
