+++ ptychodus/model/__init__.py
@@ -2,10 +2,10 @@ from .core import ModelArgs, ModelCore
 from .data import *
 from .detector import *
 from .image import *
+from .metadata import DiffractionMetadataPresenter
 from .object import *
 from .probe import *
 from .reconstructor import *
 from .scan import *
 from .tike import *
-from .velociprobe import *
 from .workflow import WorkflowPresenter, WorkflowRun
