+++ ptychodus/plugins/neXus/lynxScanFile.py
@@ -75,7 +75,3 @@ class LynxScanFileReader(ScanFileReader):
             pointList.append(ScanPoint(x_m, y_m))
 
         return SimpleScanDictionary({self.simpleName: pointList})
-
-
-def registerPlugins(registry: PluginRegistry) -> None:
-    registry.registerPlugin(LynxScanFileReader())
